+++ src/platform/file/resource_loader.h
